add first changes for python3

+++ nxswriter/Types.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 #   This file is part of nexdatas - Tango Server for NeXus data writer
 #
 #    Copyright (C) 2012-2015 DESY, Jan Kotanski <jkotan@mail.desy.de>
@@ -46,7 +46,7 @@ class Converters(object):
 class NTP(object):
 
     ## map of Python:Tango types
-    pTt = {"int": "DevLong64", "long": "DevLong64",
+    pTt = {"int": "DevLong64", "int": "DevLong64",
            "float": "DevDouble", "str": "DevString",
            "unicode": "DevString", "bool": "DevBoolean"}
 
@@ -72,9 +72,9 @@ class NTP(object):
 
     ## map of type : converting function
     convert = {"float16": float, "float32": float, "float64": float,
-               "float": float, "int64": long, "int32": int,
-               "int16": int, "int8": int, "int": int, "uint64": long,
-               "uint32": long, "uint16": int,
+               "float": float, "int64": int, "int32": int,
+               "int16": int, "int8": int, "int": int, "uint64": int,
+               "uint32": int, "uint16": int,
                "uint8": int, "uint": int, "string": str,
                "bool": Converters.toBool}
